chore: Refactor cfd offer and utilities
I cleaned up the code again. I didn't like the tightly coupled classes separated into different files. That makes it hard to find stuff. Now it's a much cleaner structure.

I introduced Tto as a prefix if we are building an opinionated widget. open for discussion.
